splits test exes into groups
